placeのリファクタリング
