+++ app/src/main/java/com/dawn/android/auth/domain/model/UserRegistration.kt
@@ -1,7 +1,7 @@
 package com.dawn.android.auth.domain.model
 
 import com.dawn.android.user.domain.model.Contact
-import com.dawn.android.user.domain.model.Place
+import com.dawn.android.place.domain.model.Place
 import com.dawn.android.user.domain.model.Sex
 import java.time.LocalDate
