+++ app/src/main/java/com/dawn/android/auth/ui/register/email/EmailRegistrationState.kt
@@ -2,11 +2,11 @@ package com.dawn.android.auth.ui.register.email
 
 import com.dawn.android.auth.domain.model.UserRegistration
 import com.dawn.android.plan.domain.model.PlaceId
-import com.dawn.android.user.domain.model.Area
-import com.dawn.android.user.domain.model.City
+import com.dawn.android.place.domain.model.Area
+import com.dawn.android.place.domain.model.City
 import com.dawn.android.user.domain.model.Contact
-import com.dawn.android.user.domain.model.Place
-import com.dawn.android.user.domain.model.Prefecture
+import com.dawn.android.place.domain.model.Place
+import com.dawn.android.place.domain.model.Prefecture
 import com.dawn.android.user.domain.model.Sex
 import java.time.LocalDate
