+++ app/src/main/java/com/dawn/android/place/domain/model/Prefecture.kt
@@ -1,4 +1,4 @@
-package com.dawn.android.user.domain.model
+package com.dawn.android.place.domain.model
 
 data class Prefecture(
     val id: Int,
