+++ app/src/main/java/com/dawn/android/plan/domain/model/PlanSchedule.kt
@@ -1,5 +1,6 @@
 package com.dawn.android.plan.domain.model
 
+import com.dawn.android.place.domain.model.Address
 import java.time.LocalTime
 
 sealed interface PlanSchedule {
